BattleArena v3.9.10.21
- fix bad import inside BattleArena.java
- fix test classes
- turned maven tests off
