+++ pom.xml
@@ -4,12 +4,13 @@
     <groupId>mc.alk</groupId>
     <artifactId>BattleArena</artifactId>
     <packaging>jar</packaging>
-    <version>3.9.10.20</version>
+    <version>3.9.10.21</version>
     <name>BattleArena</name>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.test.skip>true</maven.test.skip>
     </properties>
     <url>http://maven.apache.org</url>
     <!-- Locations for jar repositories -->
@@ -104,7 +105,7 @@
             <groupId>mc.alk</groupId>
             <artifactId>MCApi</artifactId>
             <version>2.0</version>
-            <scope>compile</scope>
+            <scope>test</scope>
         </dependency>
         <!-- End of Shaded Libraries -->
         <!-- Runtime dependencies -->
