+++ src/test/mc/alk/arena/TestConfigUpdate.java
@@ -1,4 +1,4 @@
-package test.mc.alk.arena;
+package mc.alk.arena;
 
 import java.io.BufferedReader;
 import java.io.File;
