+++ src/test/mc/alk/arena/TestJoining.java
@@ -1,13 +1,12 @@
-package test.mc.alk.arena;
+package mc.alk.arena;
 
-import java.io.File;
-import java.lang.reflect.Field;
 import junit.framework.TestCase;
 import mc.alk.arena.BattleArena;
 import mc.alk.arena.Defaults;
 import mc.alk.arena.controllers.BattleArenaController;
 import mc.alk.arena.controllers.BukkitServer;
 import mc.alk.arena.controllers.ParamController;
+import mc.alk.arena.executors.BAExecutor;
 import mc.alk.arena.objects.ArenaPlayer;
 import mc.alk.arena.objects.MatchParams;
 import mc.alk.arena.objects.arenas.Arena;
@@ -18,27 +17,33 @@ import mc.alk.arena.serializers.BAClassesSerializer;
 import mc.alk.arena.serializers.BAConfigSerializer;
 import mc.alk.arena.serializers.MessageSerializer;
 import mc.alk.arena.util.MessageUtil;
+import mc.alk.arena.util.MinMax;
 import mc.alk.mc.MCServer;
+import org.bukkit.entity.Player;
+import mc.alk.arena.objects.TestPlugin;
+import mc.alk.arena.util.Helper;
 import mc.alk.tests.testbukkit.TestBukkitPlayer;
 import mc.alk.tests.testbukkit.TestBukkitServer;
 import mc.alk.tests.testbukkit.TestMCBukkitServer;
-import org.bukkit.entity.Player;
-import test.mc.alk.arena.objects.TestPlugin;
-import test.mc.alk.arena.util.Helper;
 
-public class BATest extends TestCase {
-    TestPlugin plugin = null;
-    BattleArenaController bac;
-    private static final BAConfigSerializer baConfigSerializer = new BAConfigSerializer();
-    BattleArena ba = new BattleArena();
-    ArenaPlayer[] ap = new ArenaPlayer[10];
+import java.io.File;
+import java.lang.reflect.Field;
+
+
+@SuppressWarnings("unused")
+public class TestJoining extends TestCase{
+	TestPlugin plugin = null;
+	BattleArenaController bac;
+	private static final BAConfigSerializer baConfigSerializer = new BAConfigSerializer();
+	BattleArena ba = new BattleArena();
+	ArenaPlayer[] ap = new ArenaPlayer[10];
     public static String dir = "../arena/BattleArena";
-    public static String cdir = dir + "/test_files/competitions";
 
     @Override
-    protected void setUp() throws Exception {
+	protected void setUp() throws Exception {
         Defaults.DEBUG_MSGS = true;
         Defaults.TESTSERVER = true;
+
         System.out.println("Working Directory = " + System.getProperty("user.dir"));
         plugin = new TestPlugin();
         ArenaType.register("arena", Arena.class, plugin);
@@ -73,6 +78,7 @@ public class BATest extends TestCase {
         AnnouncementOptions.setDefaultOptions(an);
 
         mp = Helper.loadParams(dir + "/test_files/competitions/ArenaConfig.yml", plugin, "Arena");
+
         assertNotNull(mp);
 
         /// Arenas
@@ -81,24 +87,50 @@ public class BATest extends TestCase {
         as.loadArenas(plugin);
     }
 
-    public static ArenaPlayer createArenaPlayer(String name){
-        Player p1 = new TestBukkitPlayer(name);
-        return BattleArena.toArenaPlayer(p1);
+	public static ArenaPlayer createArenaPlayer(String name){
+		Player p1 = new TestBukkitPlayer(name);
+		return BattleArena.toArenaPlayer(p1);
+	}
+
+    public void testJoining() throws Exception {
+        MatchParams mp = Helper.loadParams(dir + "/test_files/competitions/ArenaConfig.yml", plugin, "Arena");
+        mp.setTeamSize(1);
+        mp.setNTeams(new MinMax(2));
+        long millisRemaining = 1000;
+        int seconds = (int) (millisRemaining / 1000);
+        int maxPlayers= mp.getMaxPlayers();
+        int minPlayers= mp.getMinPlayers();
+
+        msg(mp.getNTeams().toString()  + " - " + mp.getTeamSize().toString() +" ,min:max "+ minPlayers +" : "+maxPlayers +" , " +
+                millisRemaining + " remaining ");
     }
+
+
+    public void testFixed_stillTime2() throws Exception {
+        MatchParams mp = Helper.loadParams(dir + "/test_files/competitions/ArenaConfig.yml", plugin, "Arena");
+        mp.setTeamSize(1);
+        mp.setNTeams(new MinMax(2));
+        long millisRemaining = 1000;
+        int seconds = (int) (millisRemaining / 1000);
+        int maxPlayers= mp.getMaxPlayers();
+        int minPlayers= mp.getMinPlayers();
+
+        msg(mp.getNTeams().toString()  + " - " + mp.getTeamSize().toString() +" ,min:max "+ minPlayers +" : "+maxPlayers +" , " +
+                millisRemaining + " remaining ");
+        for (int inQ = maxPlayers; inQ < maxPlayers*2; inQ++) {
+            String m = MessageUtil.decolorChat(BAExecutor.constructMessage(mp, millisRemaining, inQ, null)).trim();
+            String e = "Your match will start when an arena is free";
+            assertEquals(e, m);
+        }
+    }
+
 
     public static void msg(String msg) {
         System.out.println(MessageUtil.decolorChat(msg));
     }
 
-    protected void delay(long millis) {
-        try {Thread.sleep(millis);}catch(Exception e){}
-    }
-    protected MatchParams loadParams(String type, String configName) throws Exception {
-        return Helper.loadParams(cdir+"/"+configName, plugin, type);
-    }
-    
-    public void testPerformed() {
-        assertTrue(true);
-    }
+	private void delay(long millis) {
+		try {Thread.sleep(millis);}catch(Exception e){}
+	}
 
 }
