--- test_files/backups/config.0.yml
@@ -1,122 +1,0 @@
-configVersion: 2.2
-
-# For help the wiki is a good source of information
-# Wiki: # https://github.com/alkarinv/BattleArena/wiki
-
-# Updates will be retrieved from the latest plugin on the bukkit site
-autoUpdate: true
-
-## What is the money on this server called, or if you use Vault the money name will come from there
-moneyName: Gold
-
-### Misc Options
-# some servers like to teleport people into the floor, you can adjust the Y offset of the teleport
-# to make them teleport higher by default, 1.0 = 1 block
-teleportYOffset: 0.0
-
-# When a player joins an arena and their inventory is stored
-# how many old inventories should be saved
-# put in 0 if you don't want this option
-numberSavedInventories: 5
-
-# which player commands should be disabled when they enter an arena
-disabledCommands: [home, spawn, payhome, warp, watch, sethome, ma]
-
-# What commands should be disabled when the player is inside of a queue, but not in a match
-disabledQueueCommands: []
-
-# which heroes skills should be disabled when they enter an arena
-disabledHeroesSkills: []
-
-# If set to true, items that are usually not stackable will be stacked when
-# a player is given items.  Examples: 64 mushroom soup, or 64 snow_ball, instead of 1 or 16
-ignoreMaxStackSize: false
-
-# if set to true, given enchanted items will not be limited to 'normal' safe enchantment levels
-# Example, most weapons are limited to sharpness 5.  if unsafeItemEnchants: true, this can be any level
-unsafeItemEnchants: false
-
-# If true if a player joins a match which has 2 arenas. 1v1 and 1v1v1v1. Then 1v1 will happen first
-# afterwards the 1v1v1v1 is guaranteed to be the next arena used.
-# if false.  if after the 1v1 is used, and the match ends, the 1v1 can be used again before the 1v1v1v1
-useArenasOnlyInOrder: false
-
-# Bukkit or Minecraft had a bug that sometimes caused players to be invisible after teleporting
-# If this is happening on your server set this to true.  
-# This option will be taken away once I have confirmed bukkit has fixed the problem
-enableInvisibleTeleportFix: true
-
-### Option sets allow you to give an easy to remember name for a group of options
-# you can add as many of your own as you want
-# storeAll: with options [storeExperience, storeGamemode, storeHealth, storeHunger, storeItems, storeHeroclass, storeMagic, clearExperience, clearInventory, deEnchant]
-# restoreAll: with options [restoreExperience, restoreGamemode, restoreHealth, restoreHunger, restoreItems, restoreHeroclass, restoreMagic, clearInventory, deEnchant]
-optionSets:
-  storeAll1:
-      options: [storeExperience, storeGamemode, storeHealth, storeHunger, storeItems, storeHeroclass, storeMagic, clearExperience, clearInventory, deEnchant]
-  restoreAll1: 
-      options: [restoreExperience, restoreGamemode, restoreHealth, restoreHunger, restoreItems, restoreHeroclass, restoreMagic, clearInventory, deEnchant]
-
-## default Options (these can be overridden by each match/event type)
-defaultOptions:
-    #### Match Options ####
-    secondsTillBegin: 6 
-    secondsTillMatch: 6 ## how long between onPrestart and onStart
-    matchTime: 120 ## How long do timed matches last, (in seconds)
-    secondsToLoot: 5 ## Time after winning to run around and collect loot
-    matchUpdateInterval: 30 ## For timed matched, how long between sending players match updates
-
-    ## when set to true when a player joins a queue the match will attempt to 
-    ## start after the forceStartTime regardless if the minimum amount of people
-    ## have joined.  Example: say 2 teams of 4 people each is needed, if after
-    ## the forceStartTime is exceeded only 2 teams of 1 person is needed to start.
-    matchEnableForceStart: true
-    matchForceStartTime: 180
-
-    ## Enable ready block (a block players can click to signify they are ready)
-    enablePlayerReadyBlock: false
-    readyBlockType: 42  ## what is the ready block (42 is iron_block)
-
-    ## How long before a player can choose another class
-    timeBetweenClassChange: 3
-
-    #### Event Options ####
-    eventCountdownTime: 180 ## How long before announcing an automated event and its start
-    eventCountdownInterval: 60 ## How often will it announce a reminder that its open and you can join
-    ## If true, when a player joins and an event that can be opened, it will be
-    ## opened and the player will join
-    allowPlayerCreation: false
-    ## When opened by a player with allowPlayerCreation, what arguments should get passed in
-    playerOpenOptions: [silent, time=180]
-  
-    #### Duel Options ####
-    allowRatedDuels: false
-    # after a player rejects a duel, how long before they can be challenged again
-    challengeInterval: 1800 # (seconds) 1800 = 30minutes
-    # allow players to challenge ops
-    allowChallengeOps: false
-
-    #### Scheduled Event Options ####
-    ## Valid options [startContinuous, startNext]
-    onServerStart: []
-    timeBetweenScheduledEvents: 30 ## seconds between end of one event, and start of another
-    announceTimeTillNextEvent: false
-  
-    #### Match/Event Announcements ####
-    ## these only affect the broadcasts to the server or channel, not the messages the fighting players receive
-    ## announce : announce these messages 
-    ## dontannounce : dont announce these messages
-    ## server : use herochat with the channel specified
-    ## herochat=<channel> : use herochat with the channel specified
-    announcements:
-        onPreStart: [ announce, server ]  ## match going to happen soon, example 'P1[p1Elo] vs P2[p2elo]'
-        onStart: [ dontannounce ]  ## match starting
-        onVictory:  [ announce, server ] ## match has been won, exmaple 'P1[p1elo] has defeated P2[p2elo]'
-        onEnterQueue:  [ dontannounce, server ] ## player1 has entered the queue for <match type>
-
-    #### Default event Announcements ####
-    eventAnnouncements:
-        onCountDownToEvent: [ announce, server ]  ## event is now open
-        onOpen: [ announce, server ]  ## event is now open
-        onStart: [ announce, server ]  ## event is starting
-        onVictory:  [ announce, server ] ## event has been won
-        onCancel:  [ announce, server ] ## event has been cancelled
